Shared helper to extract messages from various endpoint returning streams
